Extract and use app icons
These icons are dumped to `shannon_icon_cache`, and then loaded on every
subsequent launch. I was gonna originally have all the icons kept in
memory, but I couldn't get SDL_RWops to cooperate with the ZIP library
I'm using, so this is the next best approach.

This hopefully also means it should be possible to get other metadata
soon enough. :)